TIMAAT Weiterentwicklung #12 - Created and updated annotation endpoints to give options creating and receiving annotation thumbnails
* Updated annotation create endpoint to create a annotation thumbnail after creation
* Added endpoint to update the thumbnail of an annotation
* Added endpoint to receive the thumbnail of an annotation

+++ src/main/java/de/bitgilde/TIMAAT/model/FIPOP/Annotation.java
@@ -67,10 +67,10 @@ public class Annotation implements Serializable {
 	private Boolean layerVisual;
 
 	@Column(name="start_time", columnDefinition = "INT")
-	private long startTime;
+	private int startTime;
 
 	@Column(name="end_time", columnDefinition = "INT")
-	private long endTime;
+	private int endTime;
 
   @Column(name="thumbnail_position_ms")
   private Integer thumbnailPositionMs;
@@ -307,19 +307,19 @@ public class Annotation implements Serializable {
 		this.layerVisual = layerVisual;
 	}
 
-	public long getEndTime() {
+	public int getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(long endTime) {
+	public void setEndTime(int endTime) {
 		this.endTime = endTime;
 	}
 
-	public long getStartTime() {
+	public int getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(long startTime) {
+	public void setStartTime(int startTime) {
 		this.startTime = startTime;
 	}
